IMSEnabler: Application to toggle the Enhanced 4G LTE Mode setting
Somewhere in the changes made by caf to IMS the AOSP ims.apk is not
reading that Enhanced 4G LTE Mode is enabled at boot, or possibly its
a race condition.  IMS works fine if it is toggled off and back on
after network data is established. So we set up a reciever for network
change events and do the toggle when mobile data is activated.

Change-Id: I217d57cb298fed38ad6a3027b6ed81a5881097f9